rough pass migrating mlperf logging utils to a central location and adding comments
